Water Quality Analysis
A project that allows me to determine if the water is safe to drink. After removing the empty cells and the "NaN" values in the rows, I was able to perform the analysis and use the "PyCaret" library to determine which machine learning algorithm would be the best solution to this problem. It turned out that the best solution would be the Extra Trees Classifier algorithm, hence it was used for the model.